Create README.md
This is just a bunch of PHP scripts enabling the mobile application communicate with the MySQL database which is on a remote server. There are several scripts here which when called by the mobile app via Android Volley API, execute the required MySQL statements to get or store the relevant data in the database.